Move method get_engine_version to utils.py.
